add expandSingle
